rename eeprom_map header
