+++ tools/eeprom-tool/eeprom_tool.h
@@ -38,7 +38,7 @@
 #include <aditof/system.h>
 #include <aditof/eeprom_interface.h>
 #include <aditof/device_construction_data.h>
-#include "eeprom_list.h"
+#include "eeprom_map.h"
 #include "camera_eeprom_interface.h"
 
 #include <atomic>
